final_sol_aortic/requirements.txt
absl-py==0.12.0
albumentations==0.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
cycler==0.10.0
Cython==0.29.22
dataclasses==0.6
decorator==4.4.2
dill==0.3.3
docker-pycreds==0.4.0
ensemble-boxes==1.0.6
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.1
google-auth-oauthlib==0.4.4
grpcio==1.37.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.10.0
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.4
matplotlib==3.4.1
multiprocess==0.70.11.1
networkx==2.5.1
numba==0.53.1
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
pandas==1.2.3
pathtools==0.1.2
Pillow==8.2.0
promise==2.3
protobuf==3.15.8
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
sentry-sdk==1.0.0
Shapely==1.7.1
shortuuid==1.0.1
six==1.15.0
smmap==4.0.0
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
tifffile==2021.4.8
torch==1.7.0+cu101
torchaudio==0.7.0
torchvision==0.8.1+cu101
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.4
wandb==0.10.25
Werkzeug==1.0.1
yapf==0.31.0
zipp==3.4.1
